trace-cmd report: Add --ts-offset to add to timestamps of events
When merging two or more data files, if the timestamp difference between
them are known, add the --ts-offset option to let the user specify the
offset from one data file to the next. This helps merging trace data files
from hosts and guests where the timestamp offset is a known value.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>