trace-cmd record: Add --ts-offset for timestamp of events
Add an option to allow users to add a timestamp offset to all events. This
is useful when wanting to merge the resulting trace.dat file with another
trace.dat file where the timestamps are shifted by a given offset.
Specifically for virtual machines, where the host and guest have a known
offset, and the recording of the host (or guest) can use this option to
remove the need to add it when merging the trace.dat files together.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>