trace-cmd msg: Rewrote a lot of code to be more flexible in transfers
Cleaned up a lot of trace-msg.c to not be fixed in the sizes. Able to
allocate only what is needed and not to be making allocation when not
needed. Also shortened a lot of local macro names as TRACECMD doesn't need
to be specified when not exported for other files to share.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>