trace-cmd: Add page_map for entire file
Add page mapping infrastructure between pages and the file. This will allow
for the entire file to be mapped, or at least much more than just a page. If
the mmapping fails, a smaller mapping (of a power of 2 less, or half) will
be tried. This will continue till the mapping is the size of a page, and if
that fails then the entire report will fail (as it would now).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>