trace-view: Place focus back into search box after search
When searching the tree rows, after an item is found, it would be nice to
repeat the search with a keyboard instead of having to move the mouse back
to the entry and selecting again. Instead, have the focus come back to the
search entry box and allow the search to happen again with just hitting
"enter".

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>