trace-cmd: Move die() and friends out of event-utils.h and parse-utils.c
As event-utils.h and parse-utils.c are copies of the same files from the
Linux kernel source tree, move the die() funtions out of them as they no
longer exist in the kernel source.

Move the functions into trace-local.h. We need to update usage() of
kernel-shark because its prototype conflicts with the usage in
trace-local.h.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>