trace-cmd: Do not initialize profiler if it isn't being used
The profiler requires scheduler tracepoints enabled, and will bug out if
they are not. The profiler initialization will error without these
tracepoints defined. There's no reason to initialize the profiler on reading
a trace.dat file especially since it could cause that reading to fail if the
profiler required tracepoints are not there.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>