trace-cmd record: Add --quiet (-q) option
trace-cmd record will output information about the number of missed trace
events, the size of the data written to the data file from events on each
CPU, etc. When tracing programs that rely on specific output from a command,
this output can cause mis-matches that cause failures. This is particularly
true for SPEC.

When run with the --quiet (-q) option, trace-cmd record squelches this
extraneous output.

Link: http://lkml.kernel.org/r/1500488636-4569-1-git-send-email-hawkinsw@borlaugic.com

Signed-off-by: Will Hawkins <hawkinsw@borlaugic.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>