trace-cmd recorder: Always return if the read splice() does not read anything
When the splice() call to the read file descriptor returns zero or negative,
there's no reason to do a write splice() call. For some reason, the write
splice call is still called if the read splice() call returns negative and
errno is set to EAGAIN. That is wrong. It should still not call the write
splice() call.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>