trace-cmd record: Only write what was read in splice_data()
In splice_data(), it is possible that the read splice() call may not
actually read a full page_size. It should always do so, but it shouldn't be
hard coded that way. Hence, the write splice() call should not assume that
it can write a full page. Have the write splice() call only write the amount
that was read by the read splice() call.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>