trace-cmd listen: Do not kill client recorders
The client recorders are different than the tracing recorders. The clients
are connected to the network, and should run till they finish reading all
that the client has written to them. The main program should not send a
signal to kill them.

If the user does a SIGINT or SIGTERM on the main program, then it is OK for
the clients to be killed.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>