trace-cmd recorder: Add define of F_GETPIPE_SZ if not already defined
F_GETPIPE_SZ was added in Linux 2.6.35, but trace-cmd works back to 2.6.31,
and Cent OS has a kernel based on 2.6.32 where the system does not define
F_GETPIPE_SZ. This means trace-cmd will fail to build on that system.

Add a define for F_GETPIPE_SZ if it is not already defined. The system call
will fail on kernels that don't support it, and trace-cmd can already handle
(ignore) that.

Reported-by: Petar Barzakov <p.barzakov@siteground.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>