pevent: Handle new pointer processing of bprint strings
The Linux kernel printf() has some extended use cases that dereference the
pointer. This is dangerouse for tracing because the pointer that is
dereferenced can change or even be unmapped. It also causes issues when the
trace data is extracted, because user space does not have access to the
contents of the pointer even if it still exists.

To handle this, the kernel was updated to process these dereferenced
pointers at the time they are recorded, and not post processed. Now they
exist in the tracing buffer, and no dereference is needed at the time of
reading the trace.

The event parsing library needs to handle this new case.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>