trace-cmd: Remove the creating of msg out of tracecmd_msg_send()
The creation of different types is getting confusing within the
tracecmd_msg_send(). In order to split it up, we first must take the
creation of the msg out of it. This will allow us to create specific msg
types where they are used, instead of having tracecmd_msg_send() and other
functions do the multiplexing.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>