trace-cmd: Move size and cmd in tracecmd_msg into its own header struct
Move the size and cmd fields of the structure tracecmd_msg into its own
header. This will allow us to remove the individual pointers that exist in
the anonymous union of the tracecmd_msg structure into their own union. But
that will require the individual msg types size defines changing. That
change will be simplified if the header is separate.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>