trace-cmd: Use an array to map msg types and min sizes
Instead of using a switch statement, use an array to map the msg types to
the min sizes they have. If they do not have dynamically allocated data to
send, the size is simply zero, and the msg->size is used.

This incorporates a macro tuple to map the types and sizes, and will be
extended in the future for anything else that needs to be done.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>