trace-cmd: Simplify msg_read_extra()
With the extra data pointers all in the same location in tracecmd_msg
structure, and with the min_size command mapping, there's no reason to
multiplex the receiving of messages. Read the header first, then determine
by the cmd the min size to read. If more is needed, then allocate to the
generic buffer which will fill in all the other pointers.

This allows us to remove tracecmd_msg_read_extra() and use just
msg_read_extra() in its stead.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>