trace-cmd record: Fix clearing out the ctx->instance when used in for_all_instances()
The for_all_instances() macro takes the variable passed to it and uses it as
an iterator descriptor. That is, it is set to each instance, and when the
loop is complete, the variable will be NULL.

The ctx->instance is suppose to hold the main instance, and by using it as
an iterator, it will be NULL at the end. Use a local variable instance.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>