trace-cmd: Make read_proc() to return int status via OUT arg
This patch changes both the implementation and the interface of read_proc()
in trace-stack.c. First, it makes the function to read a string from the
proc file and then parse it as an integer using strtol(). Then, it makes the
function to return the integer read from the proc file using the int *status
OUT parameter, in order to make possible its return value to be used by the
caller to check if the operation succeeded.

This new implementation relaxes the external contraints the function relies
on, making it possible to be used by trace stat. The point is that 'stat'
should not fail in case there is something wrong with the stack tracer. Only
the call to die() in case the file is empty has been left in this patch: it
will be removed as well in a separate commit.

Link: http://lkml.kernel.org/r/20180116195343.12800-2-vladislav.valtchev@gmail.com

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>