trace-cmd: Remove the die() call from read_proc()
As trace-stack.c's read_proc() function is going to be used by trace-cmd
stat, we don't want it to make the program die in case something went wrong.
Therefore, this simple patch makes read_proc() to just return -1 in case the
proc file was empty or read() failed with an error, instead of using die().

Link: http://lkml.kernel.org/r/20180116195343.12800-3-vladislav.valtchev@gmail.com

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>