trace-cmd: Change the way of getting python ldflags.
Prior than this patch, Makefile detects python ldflags using a hardcoded
python command. It will cause problems if we are building against
python3 in the future when ldflags for python2 and python3 are
different. With this patch, python ldflags are detected by
corresponding python{,3}-config which will detect the right config for
python plugins.

Link: http://lkml.kernel.org/r/20180204032014.6962-3-sztsian@gmail.com

Signed-off-by: Zamir SUN (Red Hat) <sztsian@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>