trace-cmd: Move event-utils.h in lib/traceevent/include
This patch moves event-utils.h in lib/traceevent/include as part of a bigger
restructuring plan of trace-cmd's code base. In this case event-utils.h has
been moved into the private headers directory the traceevent library. In the
next steps, the source files of the traceevent lib will be moved there as well.

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>