trace-cmd: Move python-related files in python/
This patch moves all the files related with trace-cmd's python support to a
dedicated directory, as part of a bigger restructuring plan of trace-cmd's code
base. The build system has been updated to support this new directory structure
by adding a separate Makefile in python/ and by making the parent Makefile to
invoke it.

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
[ Added missing export of python_dir_SQ ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>