trace-cmd: Make the build to tell when python-dev is missing
Currently, the build system will complain when swig is missing, since that
will prevent it to build the python C modules, but in case swig is present
but python-dev is not installed, the build will just silently skip them.
This patch just makes the build to report a message when that happens.

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>