trace-cmd: Make libtraceevent buildable out-of-tree
This patch allows the libtraceevent to be buildable out-of-tree when the 'O'
variable is set at command line while invoking make.

At this stage, when the 'O' variable is set, the out-of-tree build will fail
(as it *already* does on the master branch) but with errors unrelated with
libtraceevent: with this patch the final goal of a fully working out-of-tree
build gets significantly closer.

In successive steps, the remaining Makefiles will be made to support out-of-tree
builds, until everything works.

NOTE: the regular in-tree build of all targets clearly continues to work.

Signed-off-by: Vladislav Valtchev (VMware) <vladislav.valtchev@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>