trace-cmd build: Do not execute python scripts if there is no python
I was triggering the following build messages when not having swig installed:

make: -config: Command not found

    NO_PYTHON forced: swig not installed, not compiling python plugins

make: -config: Command not found
  UPDATE                 trace_python_dir
make: -config: Command not found

This is due to the executing of $(PYTHON_VERS)-config, which would just turn
into "-config". Do not assign the PYTHON_* variables if NO_PYTHON is
defined.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>