trace-cmd: Fix bash completion on directory names
It's a pain when the bash completion of directory names adds a space after
the completion. You need to delete space, add a slash, and hit tab again.
Instead, have the completion stop after the directory name with a slash, and
then the next tab will get the files below the directory.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>