trace-filter: Change the naming convention used in trace-filter-hash
The trace-filter-hash code implements a set of hashing utilities for
filtering the trace data. Currently, only tasks are filtered with this
utility via its pid. To use this tool for filtering other items such as
events, a name change is required for the structures and functions
used. Instead of using the term "task", use the term "id", which is
more generic, and will be less confusing when using the
trace-filter-hash for filtering events and such.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>