trace-cmd: Use lookup table instead of link list for pages
When a large number of records are in use (not freed), the pages that they
are on are cached in the cpu_data. They are stored via a link list, and when
a new page is requested, that link list is searched linearly to see if it is
in the cache or not. This is increadibly insufficient as it's O(n) where n
gets large, it is very slow.

As we know the number of pages that can possibly be accessed, allocate a
table of pointers that represent each page of the file. When a new page is
requested, the table is examined to see if it is already allocated, and if
it is, it is returned, otherwise it is allocated and added to the lookup
table.

When a page is no longer referenced, it is freed and the lookup table for
that page location is cleared.

This changes the lookup time from O(n) to O(1) with the sacrifice of a bit
more memory usage. But this proved to be worth it in the new KernelShark
development.

Reported-by: "Yordan Karadzhov (VMware)" <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>