kernel-shark-qt: Make helper functions recognize data type to improve performance
Add a type enumerator and change rec_list into a union such that the helper
functions have more information about the type of data they are processing.
This will allow for the helper functions to be optimized such that they
perform better, and we can remove double allocations.

Some of the functionality of kshark_load_data_records() and
kshark_load_data_entries() have been moved into the helper functions with a
switch statement based on the type variable to know what to do with the
data.

The rec_list structure's entry element has been converted from a pointer,
that needs to be allocated, to a structure itself, such that it can be
allocated to store the entry data without having to allocate extra data.

The next field of the kshark_entry has been moved to the top of the
structure to match the rec_list next entry, so that they both have the link
list pointer as the first entry, and it does not matter if the rec_list is
used for kshark_entry's or for records.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>