kernel-shark-qt: Add I/O for configuration data.
Add to the C API of KernelShark instruments for importing/exporting
configuration settings to/from Json files. A wrapper is added in order
to abstract out completely the use of Json. This can be useful in the
future if we want to support multiple formats (YAML for example).

This version of the patch contains a number of improvements suggested
by Steven Rostedt. Thanks Steven!

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>