kernel-shark-qt: Change the color convention of the task graphs
So far the Task graphs have used the same color convention as the
CPU graphs. Namely the vertical tick of the record was plotted by
using the CPU color and the horizontal strap was plotted by using
the task color. This patch reverts the color convention and now the
tick has the color of the task and the strap has the color of the
CPU. This change requires significant modification of the logic
used to create and draw the graphs.

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>