kernel-shark-qt: Improve the plotting logic of the Sched event plugin
This patch aims to make the plotting logic of the Sched event plugin
more robust and easy to understand. It also provides a proper
processing of the case when we have multiple Sched events in one bin.
The red box showing the time the task was preempted is open only if
the prev_state of the sched_switch event is "R".

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>