kernel-shark-qt: Update Sched Events plugin
This patch updates the Sched Events plugin to use the recently
introduced "Missed events" entries and tot_count field of the
Visualization model descriptor in its plotting logic. We need this
update because the sched plugin would get confused because of missed
events, where we could have missed events hiding wakeups, and only
have the sched switch, or a wakeup with a missing sched switch would
break the plotting of the latency.

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>