kernel-shark-qt: Optimize the search in the text data
This patch aims to optimize the search by doing two small
modifications:

It changes the type of the matching function used when searching.
This will eliminate the call of the QString copy constructor.

It also adds an intermediate step in the conversion of the tracing
data into the format expected by the QTableView widget (QVariant).
By having this intermediate step we avoid the unnecessary conversion
QString -> QVariant -> QString when searching.

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>