kernel-shark-qt: Don't use collection when searching for sched_wakeup
This patch modifies the way the wake up latency is plotted by
the Sched events plugin. The sched_wakeup event does not belong to
the task it is waking, hence we cannot use the task's collection
to search for it. It was partially working so far because in most
of the cases the sched_wakeup event happened to be inside the
"margins" of the collection. Note that if we do not use collection
when searching, the time complexity of the plugin becomes linear.
However, this should not be a problem, because we have an explicit
cut which disables plotting if the number of entries shown by
the graph is bigger than PLUGIN_MAX_ENTRIES (10 000).

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>