kernel-shark-qt: Add a method for adding a new collection to a list
So far we have only one method that creates a new Data Collection.
It is kshark_register_data_collection(). The problem with this method
is that it automatically adds the new collection to the list of
collections, owned by the context of the session. This patch adds
a more generic method, which can be used to add new data collection
to a given list. The method will be used by the Sched event plugin,
which will keep its own list of collections, relevant only for the
plugin.

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>