trace-cmd: Fix record --date flag when sending tracing data to a listener
Currently the `trace-cmd record` --date is not taken into account when tracing
data is sent to a remote host with the -N flag.

This patch fixes this by the writing output buffer options from the recording
side instead of on the listener side.

We also provide backward compatibility by falling back to previous behavior when
we're not using protocol v3.

Signed-off-by: Slavomir Kaslev <kaslevs@vmware.com>
[ Changed compare to version "< 3" and added bisectability comment ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>