kernel-shark-qt: Implement State machine for searching in the data
This patch was originally motivated by the request to make it
possible to continue searching from the place when the search
has been stopped by the user. However because the implementation
of this feature made the original code quite messy (ugly), I
decided to reorganize the way the searching is controlled and
add a KsSearchFSM class that implements a Finite-State Machine
by following the "State" design pattern. KsSearchFSM class now
encapsulates the searching logic.

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>